feat(esp-box): Update implementation * Break up implementation into different source files * Update audio interface to allow configuration of audio task
